Split maxItems normalization internals into focused files
